chore: unblock capn pre-push — clippy / docs / glob / test cleanup
Pre-existing main-branch failures that were blocking `capn pre-push`:

daw-proto:
  - Add #[allow(ambiguous_glob_reexports)] crate-level (the per-RPC
    modules each re-export Service/Dispatcher/descriptor/layer/serve;
    only the qualified paths `daw_proto::marker::Service` etc are used
    by callers — the bare-name binding is harmless).
  - Convert `pub use action_registry::*` / `pub use window_manager::*`
    in lib.rs to explicit re-exports matching the pattern already used
    for marker/region/track/take/tempo_map/ext_state.
  - Remove redundant `pub use ext_state::*` (lines 93-95 already do
    the explicit version).
  - Rename `DockHostService` → `DockHosting` in dock_host_mock.rs
    (the trait was renamed; mock didn't follow).
  - LayoutPlacement: derive Default with #[default] on Hidden variant.
  - Marker: tag the manual Default impl with #[allow(derivable_impls)]
    + comment explaining Position::start() == Position::default() but
    the explicit constructor reads better.

daw-link:
  - Collapse the nested `if let / if / if let` chain at lib.rs:710
    using let-else chain syntax.

daw-audio-sync:
  - Strip broken intra-doc-links to PeerTable and ArcSwap (the linked
    items live in private modules; plain backticks instead).

dawfile-protools:
  - examples/dump_102d.rs: replace `ContentType::TrackPerTrackState`
    (never existed) with raw `content_type_raw == 0x102d` check.

dawfile-reaper:
  - Fix two unit tests (`test_indentation_nesting`,
    `test_item_with_wave_source`) that asserted `TAKE SEL` would appear
    in serialized output. The serializer documents (and implements)
    "omit TAKE/TAKE SEL marker for take #0" — REAPER reads take 0's
    properties directly off the ITEM block. Tests now match the
    documented behavior.

.config/nextest.toml:
  - Add per-test override giving `round_trip_every_fixture` 60s before
    termination (default profile) / 120s (ci profile). The test spawns
    the round_trip example as a subprocess for every PT fixture
    (~18 files); debug-mode parse + process startup easily exceeds the
    default 15s budget. Test passes cleanly in 17s release / ~40s debug.

cargo clippy --fix:
  - Mechanical cleanup of unused imports, collapsible-if blocks,
    inline-format-args, and similar across ~70 files. Behavior
    unchanged; tests all green (workspace lib+bins: 0 failures).

+++ crates/daw-reaper/examples/find_grown_block.rs
@@ -6,7 +6,7 @@ use dawfile_protools::raw_block::RawBlock;
 fn flatten(blocks: &[RawBlock], out: &mut Vec<(u16, usize, u32, usize)>) {
     fn walk(blocks: &[RawBlock], depth: usize, out: &mut Vec<(u16, usize, u32, usize)>) {
         for b in blocks {
-            out.push((b.content_type_raw, b.start, b.block_size as u32, depth));
+            out.push((b.content_type_raw, b.start, b.block_size, depth));
             walk(&b.children, depth + 1, out);
         }
     }
